feat: refactor components to render them dynamically
