+++ src/routes/__root.tsx
@@ -1,6 +1,6 @@
 import { createRootRoute, Outlet } from "@tanstack/react-router";
 import { SidebarProvider, SidebarInset } from "@rhino-ui/ui";
-import { AppSidebar } from "@/components/sidebar/app-sidebar";
+import { AppSidebar } from "@/components/sidebar/AppSidebar";
 
 import { TanStackRouterDevtools } from "@tanstack/router-devtools";
