+++ src/types/components/FieldTypes.ts
@@ -1,27 +1,3 @@
-export type RequestType =
-	| "GET"
-	| "POST"
-	| "PUT"
-	| "DELETE"
-	| "PATCH"
-	| "HEAD"
-	| "OPTIONS"
-	| "CONNECT"
-	| "TRACE";
-
-export interface RequestItem {
-	title: string;
-	url: string;
-	requestType: RequestType;
-}
-
-export interface Request {
-	title: string;
-	url: string;
-	isActive?: boolean;
-	items?: RequestItem[];
-}
-
 export interface BaseField {
 	name: string;
 	type:
@@ -39,11 +15,12 @@ export interface BaseField {
 	minimum?: number;
 	maximum?: number;
 	pattern?: string;
+	defaultValue?: string | number | boolean;
+	options?: string[];
 }
 
 export interface PrimitiveField extends BaseField {
 	type: "string" | "integer" | "boolean" | "number" | string;
-	defaultValue?: string | number | boolean;
 }
 
 export interface ArrayField extends BaseField {
