+++ src/types/sidebar.ts
@@ -1,4 +1,4 @@
-import type { Request } from "./requests";
+import type { Request } from "./api/requests";
 
 export interface SidebarCollection {
 	collectionName: string;
